Move translatable component test to the correct test class
