+++ chat/src/test/java/net/md_5/bungee/api/chat/ComponentsTest.java
@@ -565,29 +565,6 @@ public class ComponentsTest
     }
 
     @Test
-    public void testTranslateComponent()
-    {
-        TranslatableComponent item = new TranslatableComponent( "item.swordGold.name" );
-        item.setColor( ChatColor.AQUA );
-        TranslatableComponent translatableComponent = new TranslatableComponent( "commands.give.success",
-                item, "5",
-                "thinkofdeath" );
-
-        assertEquals( "Given Golden Sword * 5 to thinkofdeath", translatableComponent.toPlainText() );
-        assertEquals( ChatColor.WHITE + "Given " + ChatColor.AQUA + "Golden Sword" + ChatColor.WHITE
-                + " * " + ChatColor.WHITE + "5" + ChatColor.WHITE + " to " + ChatColor.WHITE + "thinkofdeath",
-                translatableComponent.toLegacyText() );
-
-        TranslatableComponent positional = new TranslatableComponent( "book.pageIndicator", "5", "50" );
-
-        assertEquals( "Page 5 of 50", positional.toPlainText() );
-        assertEquals( ChatColor.WHITE + "Page " + ChatColor.WHITE + "5" + ChatColor.WHITE + " of " + ChatColor.WHITE + "50", positional.toLegacyText() );
-
-        TranslatableComponent one_four_two = new TranslatableComponent( "filled_map.buried_treasure" );
-        assertEquals( "Buried Treasure Map", one_four_two.toPlainText() );
-    }
-
-    @Test
     public void testBuilderCreate()
     {
         testBuilder(
