+++ chat/src/test/java/net/md_5/bungee/api/chat/TranslatableComponentTest.java
@@ -1,6 +1,7 @@
 package net.md_5.bungee.api.chat;
 
 import static org.junit.jupiter.api.Assertions.*;
+import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.chat.ComponentSerializer;
 import org.junit.jupiter.api.Test;
 
@@ -25,4 +26,27 @@ public class TranslatableComponentTest
         assertEquals( "Test string with a placeholder", BaseComponent.toPlainText( baseComponents ) );
         assertEquals( "§fTest string with §fa§f placeholder", BaseComponent.toLegacyText( baseComponents ) );
     }
+
+    @Test
+    public void testTranslateComponent()
+    {
+        TranslatableComponent item = new TranslatableComponent( "item.swordGold.name" );
+        item.setColor( ChatColor.AQUA );
+        TranslatableComponent translatableComponent = new TranslatableComponent( "commands.give.success",
+                item, "5", "thinkofdeath" );
+
+        assertEquals( "Given Golden Sword * 5 to thinkofdeath", translatableComponent.toPlainText() );
+        assertEquals( ChatColor.WHITE + "Given " + ChatColor.AQUA + "Golden Sword" + ChatColor.WHITE
+                        + " * " + ChatColor.WHITE + "5" + ChatColor.WHITE + " to " + ChatColor.WHITE + "thinkofdeath",
+                translatableComponent.toLegacyText() );
+
+        TranslatableComponent positional = new TranslatableComponent( "book.pageIndicator", "5", "50" );
+
+        assertEquals( "Page 5 of 50", positional.toPlainText() );
+        assertEquals( ChatColor.WHITE + "Page " + ChatColor.WHITE + "5" + ChatColor.WHITE + " of " + ChatColor.WHITE + "50", positional.toLegacyText() );
+
+        TranslatableComponent one_four_two = new TranslatableComponent( "filled_map.buried_treasure" );
+        assertEquals( "Buried Treasure Map", one_four_two.toPlainText() );
+    }
+
 }
